bajacombi error messages fixed

+++ app/Http/Controllers/CombiController.php
@@ -110,13 +110,21 @@ class CombiController extends Controller
 
         return redirect()->to(route('combi.info', ['combi' => $combi]))->with('combimodificado', 'open');
     }
+
+
     public static function destroy(Combi $combi)
     {
 
         if(isSet($combi->chofer)){
 
-            return redirect()->to(route('combi.info', ['combi' => $combi]))-> with('bajaerronea',$combi);
+            return redirect()->to(route('combi.info', ['combi' => $combi]))-> with('tienechofer',$combi);
         }
+
+        if(isSet($combi->ruta)){
+            return redirect()->to(route('combi.info', ['combi' => $combi]))-> with('tieneruta',$combi);
+        }
+
+
         DB::table('combis')->whereId($combi->id)->delete();
 
         return redirect()->to(route('combi.listar'))-> with('combieliminada',$combi);
